pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["dopefresh <dopefresh4000@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.5"
djangorestframework = "^3.12.4"
environs = "^9.3.2"
django-cors-headers = "^3.7.0"
psycopg2 = "^2.9.1"
djangorestframework-simplejwt = "^4.7.2"
loguru = "^0.5.3"
drf-spectacular = "^0.17.3"
gunicorn = "^20.1.0"
Pillow = "^8.3.1"
flake8 = "^3.9.2"
factory-boy = "^3.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
